fix poesy tensor shape
