+++ py4cast/datasets/poesy/__init__.py
@@ -94,7 +94,6 @@ class PoesyAccessor(DataAccessor):
         ds_name: str,
         param: WeatherParam,
         timestamps: Timestamps,
-        term: np.array,
         member: int,
         file_format: str = "npy",
     ) -> np.array:
@@ -102,13 +101,18 @@ class PoesyAccessor(DataAccessor):
         date : Date of file.
         term : Position of leadtimes in file.
         """
-        data_array = np.load(self.get_filepath(ds_name, param, timestamps.datetime), mmap_mode="r")
-        return data_array[
+        data_array = np.load(
+            self.get_filepath(ds_name, param, timestamps.datetime), mmap_mode="r"
+        )
+
+        arr = data_array[
             param.grid.subdomain[0] : param.grid.subdomain[1],
             param.grid.subdomain[2] : param.grid.subdomain[3],
-            (timestamps.term / dt.timedelta(hours=1)).astype(int) - 1,
+            (timestamps.terms / dt.timedelta(hours=1)).astype(int) - 1,
             member,
-        ]
+        ].transpose([2, 0, 1])
+
+        return np.expand_dims(arr, -1)
 
     def exists(
         self,
@@ -123,7 +127,7 @@ class PoesyAccessor(DataAccessor):
             return False
         return True
 
-    def valid_timestamp(self, n_inputs: int, timestamps: Timestamps) -> bool:
+    def valid_timestamp(n_inputs: int, timestamps: Timestamps) -> bool:
         """
         Verification function called after the creation of each timestamps.
         Check if computed terms respect the dataset convention.
@@ -138,14 +142,6 @@ class PoesyAccessor(DataAccessor):
             ):
                 return False
         return True
-
-class InferSample(Sample):
-    """
-    Sample dedicated to inference. No outputs terms, only inputs.
-    """
-
-    def __post_init__(self):
-        self.terms = self.input_terms
 
 
 class PoesyDataset(DatasetABC):
@@ -162,130 +158,3 @@ class PoesyDataset(DatasetABC):
 
     def __len__(self):
         return len(self.sample_list)
-
-
-# class InferPoesyDataset(PoesyDataset):
-#     """
-#     Inherite from the PoesyDataset class.
-#     This class is used for inference, the class overrides methods sample_list and from_json.
-#     """
-
-#     def __init__(self, *args, **kwargs):
-#         super().__init__(*args, **kwargs)
-
-#     @cached_property
-#     def sample_list(self):
-#         """
-#         Create a list of sample from information.
-#         Outputs terms are computed from the number of prediction steps in argument.
-#         """
-#         print("Start forming samples")
-#         terms = list(
-#             np.arange(
-#                 METADATA["TERMS"]["start"],
-#                 METADATA["TERMS"]["end"],
-#                 METADATA["TERMS"]["timestep"],
-#             )
-#         )
-
-#         num_total_steps = self.settings.num_input_steps + self.settings.num_pred_steps
-#         sample_by_date = len(terms) // num_total_steps
-#         samples = []
-#         number = 0
-#         for date in self.period.date_list:
-#             for member in self.settings.members:
-#                 for sample in range(0, sample_by_date):
-#                     input_terms = terms[
-#                         sample * num_total_steps : sample * num_total_steps
-#                         + self.settings.num_input_steps
-#                     ]
-
-#                     output_terms = [
-#                         input_terms[-1] + METADATA["TERMS"]["timestep"] * (step + 1)
-#                         for step in range(self.settings.num_pred_steps)
-#                     ]
-
-#                     samp = InferSample(
-#                         date=date,
-#                         member=member,
-#                         settings=self.settings,
-#                         input_terms=input_terms,
-#                         output_terms=output_terms,
-#                     )
-
-#                     if samp.is_valid():
-#                         samples.append(samp)
-#                         number += 1
-#         print(f"All {len(samples)} samples are now defined")
-
-#         return samples
-
-#     @classmethod
-#     def from_json(
-#         cls,
-#         fname: Path,
-#         num_input_steps: int,
-#         num_pred_steps_train: int,
-#         num_pred_steps_val_tests: int,
-#         config_override: Union[Dict, None] = None,
-#     ) -> Tuple[None, None, "InferPoesyDataset"]:
-#         """
-#         Return 1 InferPoesyDataset.
-#         Override configuration file if needed.
-#         """
-
-#         with open(fname, "r") as fp:
-#             conf = json.load(fp)
-#             if config_override is not None:
-#                 conf = merge_dicts(conf, config_override)
-#                 print(conf["periods"]["test"])
-#         conf["grid"]["load_grid_info_func"] = load_grid_info
-#         grid = Grid(**conf["grid"])
-#         param_list = get_param_list(conf, grid, load_param_info, get_weight_per_level)
-#         inference_period = Period(**conf["periods"]["test"], name="infer")
-
-#         ds = InferPoesyDataset(
-#             grid,
-#             inference_period,
-#             param_list,
-#             SamplePreprocSettings(
-#                 num_pred_steps=0,
-#                 num_input_steps=num_input_steps,
-#                 members=conf["members"],
-#                 **conf["settings"],
-#             ),
-#         )
-
-#         return None, None, ds
-
-
-if __name__ == "__main__":
-    path_config = "config/datasets/poesy.json"
-
-    parser = ArgumentParser(description="Prepare Poesy dataset and test loading speed.")
-    parser.add_argument(
-        "--path_config",
-        default=path_config,
-        type=Path,
-        help="Configuration file for the dataset.",
-    )
-    parser.add_argument(
-        "--n_iter",
-        default=10,
-        type=int,
-        help="Number of samples to test loading speed.",
-    )
-    args = parser.parse_args()
-
-    PoesyDataset.prepare(args.path_config)
-
-    print("Dataset info : ")
-    train_ds, _, _ = PoesyDataset.from_json(args.path_config, 2, 3, 3)
-    train_ds.dataset_info.summary()
-
-    print("Test __get_item__")
-    print("Len dataset : ", len(train_ds))
-
-    print("First Item description :")
-    data_iter = iter(train_ds.torch_dataloader())
-    print(next(data_iter))
